Use WS config file directly
This changes the command-line options.

Formerly the script used a symbolic link to this file, so that the custom keymapper would not be created in teh WS directory, Not the script uses a dosbox setting to specify a custom location for the mapper file, so it doesn't have to reside in the same directory as the config file.